Removed all ref of input files (#39)
